+++ tests/ClientTest.php
@@ -87,7 +87,7 @@ class ClientTest extends \PHPUnit_Framework_TestCase
 
     public function testCanMergeOnBaseUriWithRequest()
     {
-        $mock = new MockHandler([new Response()]);
+        $mock = new MockHandler([new Response(), new Response()]);
         $client = new Client([
             'handler'  => $mock,
             'base_uri' => 'http://foo.com/bar/'
@@ -96,6 +96,13 @@ class ClientTest extends \PHPUnit_Framework_TestCase
         $this->assertEquals(
             'http://foo.com/bar/baz',
             (string) $mock->getLastRequest()->getUri()
+        );
+
+        $client->request('GET', new Uri('baz'), ['base_uri' => 'http://example.com/foo/']);
+        $this->assertEquals(
+            'http://example.com/foo/baz',
+            (string) $mock->getLastRequest()->getUri(),
+            'Can overwrite the base_uri through the request options'
         );
     }
 
@@ -586,7 +593,7 @@ class ClientTest extends \PHPUnit_Framework_TestCase
 
     public function testProperlyBuildsQuery()
     {
-        $mock = new MockHandler([new Response(200)]);
+        $mock = new MockHandler([new Response()]);
         $client = new Client(['handler' => $mock]);
         $request = new Request('PUT', 'http://foo.com');
         $client->send($request, ['query' => ['foo' => 'bar', 'john' => 'doe']]);
@@ -595,8 +602,8 @@ class ClientTest extends \PHPUnit_Framework_TestCase
 
     public function testSendSendsWithIpAddressAndPortAndHostHeaderInRequestTheHostShouldBePreserved()
     {
-        $mockHandler = new MockHandler([new Response(200)]);
-        $client = new Client(['base_uri' => '127.0.0.1:8585', 'handler' => $mockHandler]);
+        $mockHandler = new MockHandler([new Response()]);
+        $client = new Client(['base_uri' => 'http://127.0.0.1:8585', 'handler' => $mockHandler]);
         $request = new Request('GET', '/test', ['Host'=>'foo.com']);
 
         $client->send($request);
@@ -606,7 +613,7 @@ class ClientTest extends \PHPUnit_Framework_TestCase
 
     public function testSendSendsWithDomainAndHostHeaderInRequestTheHostShouldBePreserved()
     {
-        $mockHandler = new MockHandler([new Response(200)]);
+        $mockHandler = new MockHandler([new Response()]);
         $client = new Client(['base_uri' => 'http://foo2.com', 'handler' => $mockHandler]);
         $request = new Request('GET', '/test', ['Host'=>'foo.com']);
 
@@ -615,4 +622,35 @@ class ClientTest extends \PHPUnit_Framework_TestCase
         $this->assertEquals('foo.com', $mockHandler->getLastRequest()->getHeader('Host')[0]);
     }
 
+    /**
+     * @expectedException \InvalidArgumentException
+     */
+    public function testValidatesSink()
+    {
+        $mockHandler = new MockHandler([new Response()]);
+        $client = new Client(['handler' => $mockHandler]);
+        $client->get('http://test.com', ['sink' => true]);
+    }
+
+    public function testHttpDefaultSchemeIfUriHasNone()
+    {
+        $mockHandler = new MockHandler([new Response()]);
+        $client = new Client(['handler' => $mockHandler]);
+
+        $client->request('GET', '//example.org/test');
+
+        $this->assertSame('http://example.org/test', (string) $mockHandler->getLastRequest()->getUri());
+    }
+
+    public function testOnlyAddSchemeWhenHostIsPresent()
+    {
+        $mockHandler = new MockHandler([new Response()]);
+        $client = new Client(['handler'  => $mockHandler]);
+
+        $client->request('GET', 'baz');
+        $this->assertSame(
+            'baz',
+            (string) $mockHandler->getLastRequest()->getUri()
+        );
+    }
 }